+++ .gitignore
@@ -105,6 +105,7 @@ celerybeat.pid
 # Environments
 .env
 .venv
+venv3.9
 env/
 venv/
 ENV/
